requirements.txt
fastapi
uvicorn
langchain
langchain_community
huggingface_hub
langchain-openai 
langchainhub